API

MINDFulMakie.ibnplot! — Method

intentplot(ibn::IBN)
intentplot!(ax, ibn::IBN)

Creates a graph plot of the ibn.

Attributes

show_routers = false: show labels for nodes

show_links = false: show labels for links.

subnetwork_view = false: show node labels with the controller view indexing.

intentidx = nothing: highligh the passed in intent.

MINDFulMakie.ibnplot — Method

intentplot(ibn::IBN)
intentplot!(ax, ibn::IBN)

Creates a graph plot of the ibn.

Attributes

show_routers = false: show labels for nodes

show_links = false: show labels for links.

subnetwork_view = false: show node labels with the controller view indexing.

intentidx = nothing: highligh the passed in intent.

MINDFulMakie.intentplot! — Method

intentplot(ibn::IBN, idx::Integer)
intentplot!(ax, ibn::IBN, idx::Integer)

Creates a tree plot of the intent Directed Acyclic Graph (DAG) idx.

Attributes

interdomain=false: reach out to all domains (might malfunction)

show_state=true: shows the state of each intent DAG node

MINDFulMakie.intentplot — Method

intentplot(ibn::IBN, idx::Integer)
intentplot!(ax, ibn::IBN, idx::Integer)

Creates a tree plot of the intent Directed Acyclic Graph (DAG) idx.

Attributes

interdomain=false: reach out to all domains (might malfunction)

show_state=true: shows the state of each intent DAG node

MINDFulMakie.netgraphplot! — Method

netgraphplot(ng::NetsedGraph)
netgraphplot!(ax, ng::NetsedGraph)

Creates a graph plot of the ng represented as a NestedGraph.

Attributes

show_routers = false: show labels for nodes

show_links = false: show labels for links.

color_paths = nothing: color given paths.

You must pass in Vector{Vector{Integer}} where each nested Vector is a path. Each path is being colored with a distinct color.

color_edges = nothing: color given edges.

Pass in Vector{Vector{Edge}}. Each nested Vector{Edge} gets a distinct color.

pure_colors = nothing: enter a Vector of colors to use for edges/paths.

circle_nodes = nothing: choose which nodes to circle.

Pass in Vector{Vector{Int}}. Each nested Vector{Int} gets the same color.

MINDFulMakie.netgraphplot — Method

netgraphplot(ng::NetsedGraph)
netgraphplot!(ax, ng::NetsedGraph)

Creates a graph plot of the ng represented as a NestedGraph.

Attributes

show_routers = false: show labels for nodes

show_links = false: show labels for links.

color_paths = nothing: color given paths.

You must pass in Vector{Vector{Integer}} where each nested Vector is a path. Each path is being colored with a distinct color.

color_edges = nothing: color given edges.

Pass in Vector{Vector{Edge}}. Each nested Vector{Edge} gets a distinct color.

pure_colors = nothing: enter a Vector of colors to use for edges/paths.

circle_nodes = nothing: choose which nodes to circle.

Pass in Vector{Vector{Int}}. Each nested Vector{Int} gets the same color.

Index

MINDFulMakie.ibnplot

MINDFulMakie.ibnplot!

MINDFulMakie.intentplot

MINDFulMakie.intentplot!

MINDFulMakie.netgraphplot

MINDFulMakie.netgraphplot!